requirements.txt
fastapi
uvicorn
faiss-cpu
sqlalchemy
pydantic
python-multipart
openai
sentence-transformers
